requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.4.0
black==25.11.0
click==8.3.1
colorama==0.4.6
iniconfig==2.3.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.1
pytest-asyncio==1.3.0
pytokens==0.3.0
websockets==15.0.1
